Renamed file for readability (See desc.)
Hey there! I changed your file name from "1.py" to "001.py" as a suggestion for the rest of your files. Currently with the setup you have your problems get split up a lot like this:
- 1.py
- 11.py
- 12.py
- 123.py
- ..
- 2.py
- 20.py 
And so on. I renamed this file to 001.py as a suggestion that you perhaps you switch to a three number categorization setup so that if other people look at your solutions (or yourself) they have can find what they are looking for quicker and looks nicer just to look at. :)

So essentially your new list would like like:
- 001.py
- 002.py
- 003.py
- .. 
- 010.py
- ..
- 123.py

And so on. 

Anyway, feel free to deny the merge and not change the names, I just think it's a quick and easy way to give your repo a cleaner and user-friendly look so I figured I'd stop by and leave a recommendation, have a nice day. ^^